+++ packages/jetbrains-plugin/src/test/resources/project-fixtures/basic-java-project-with-mongodb/src/main/java/alt/mongodb/springcriteria/BookRepository.java
@@ -1,6 +1,7 @@
 package alt.mongodb.springcriteria;
 
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.aggregation.Aggregation;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.List;
@@ -19,5 +20,17 @@ class BookRepository {
 
     public List<Book> allReleasedBooks() {
         return template.query(Book.class).matching(where("released").is(true)).all();
+    }
+
+    public List<Book> allReleasedBooksAgg() {
+        return template.aggregate(
+            Aggregation.newAggregation(
+                Aggregation.match(
+                    where("released").is(true)
+                )
+            ),
+            Book.class,
+            Book.class
+        ).getMappedResults();
     }
 }